EPMGCIP-163: Add new hook to track mobile view layout (sm breakpoint), update hierarchy for Header sub components
