+++ src/components/organisms/Header/Header.module.scss
@@ -9,7 +9,7 @@
   border-bottom: 1px solid $mainBorderColor;
   display: flex;
   flex-direction: row;
-  justify-content: flex-start;
+  justify-content: space-between;
   align-items: center;
   height: 60px;
 
@@ -32,7 +32,10 @@
   height: 100%;
   display: flex;
   flex-direction: row;
+  justify-content: flex-start;
+  align-items: center;
   padding: 0;
+  margin: 0 5%;
 }
 
 .link {
@@ -71,14 +74,6 @@
     border-top-left-radius: 5px;
     border-top-right-radius: 5px;
   }
-}
-
-.burgerMenuIcon {
-  margin-left: 15px;
-}
-
-.desktopHeaderContent {
-  height: 100%;
 }
 
 .subLinksDropdown {
