+++ src/hooks/use-mobile-view/index.ts
@@ -1,0 +1,1 @@
+export { useMobileView } from './useMobileView';
