+++ src/hooks/use-mobile-view/useMobileView.ts
@@ -1,0 +1,8 @@
+import { em } from '@mantine/core';
+import { useMediaQuery } from '@mantine/hooks';
+
+const smMobileViewMaxWidth = 768;
+
+export const useMobileView = () => {
+  return useMediaQuery(`(max-width: ${em(smMobileViewMaxWidth)})`);
+};
